Added shortcuts from operation page and from Me plus refactoring
Tapping "Mnsk" will on operation page will redirect to list of Personnel
Tapping "Enheter" on operation page will redirect to list of Units
Tapping "Meg" in map will redirect you to My page

Some refactoring of MyLocationLayer
